Extract rails controller logic into its own class
