+++ spec/mj/alternative_file/resolvers/ruby/rails_resolver_spec.rb
@@ -28,34 +28,4 @@ RSpec.describe Mj::AlternativeFile::Resolvers::Ruby::RailsResolver do
 
     expect(result).to include(create_candidate("test/models/users/profile_test.rb", "minitest"))
   end
-
-  context "when handling rails controllers" do
-    let(:controller) { "app/controllers/foos/bars_controller.rb" }
-    let(:integration_test) { "test/integration/foos/bars_controller_test.rb" }
-    let(:integration_spec) { "spec/integration/foos/bars_controller_spec.rb" }
-
-    it "resolves controller integration test" do
-      result = resolve(controller)
-
-      expect(result).to include(create_candidate(integration_test, "integration_test"))
-    end
-
-    it "resolves controller integration spec" do
-      result = resolve(controller)
-
-      expect(result).to include(create_candidate(integration_spec, "integration_spec"))
-    end
-
-    it "resolves controller from integration test" do
-      result = resolve(integration_test)
-
-      expect(result).to include(create_candidate(controller, "controller"))
-    end
-
-    it "resolves controller from integration spec" do
-      result = resolve(integration_spec)
-
-      expect(result).to include(create_candidate(controller, "controller"))
-    end
-  end
 end
